+++ sdk/core/azure-core/test/ut/transport_adapter_base.cpp
@@ -522,7 +522,7 @@ namespace Azure { namespace Core { namespace Test {
       int64_t size,
       std::string expectedBody)
   {
-    auto body = response.GetBodyStream();
+    auto body = response.ExtractBodyStream();
     EXPECT_EQ(body, nullptr);
     std::vector<uint8_t> bodyVector = response.GetBody();
     int64_t bodySize = bodyVector.size();
@@ -544,7 +544,7 @@ namespace Azure { namespace Core { namespace Test {
       int64_t size,
       std::string expectedBody)
   {
-    auto body = response.GetBodyStream();
+    auto body = response.ExtractBodyStream();
     EXPECT_NE(body, nullptr);
 
     std::vector<uint8_t> bodyVector